% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{all_raw_data}
\alias{all_raw_data}
\title{This fuction writes a .tsv file after merging all the drugEraM2.rda files together.
This gives one .tsv file that contains all the data thus far gathered.
db names are appended to the left of the data set.}
\usage{
all_raw_data(item = character(),
  data_folder = "//lhcdevfiler/cgsb-minf/yohan/OHDSITrends-pre/data",
  dest_folder = ".tsv files")
}
\description{
This fuction writes a .tsv file after merging all the drugEraM2.rda files together.
This gives one .tsv file that contains all the data thus far gathered.
db names are appended to the left of the data set.
}
