% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{combine_items}
\alias{combine_items}
\title{This function takes the classified items (full_cids), appends a column
to the left, indicating which database the items are from, merges the items,
and then prints a single csv file for all the items}
\usage{
combine_items(dom, DB = c("one", "two", "three"), dir = NULL)
}
\arguments{
\item{dom}{Character vector of domains (\code{c('drug', 'proc', 'cond')})}

\item{DB}{Character of the DBs (\code{c('one', 'two', 'three')})}

\item{dir}{file path to where the clasified items live.
this directory is also where the output .csv file will go}
}
\value{
None
}
\description{
This function takes the classified items (full_cids), appends a column
to the left, indicating which database the items are from, merges the items,
and then prints a single csv file for all the items
}
\note{
No warning messages are thrown if the input names are bad.

csv file name is all_{domain}_items
}
