% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{count_filter}
\alias{count_filter}
\title{@note This function filters age-stratified medical events that have a minimum occurance frequency
in at least one year}
\usage{
count_filter(d, threshold = 1)
}
\arguments{
\item{d}{data to filter}

\item{threshold}{minimum occurance frequency in people per thousand; 1 is default}
}
\value{
itemsA are all the age-stratified medical events that pass the threshold

itemsB are all the age-stratified medical events that fail the threshold.
}
\description{
@note This function filters age-stratified medical events that have a minimum occurance frequency
in at least one year
}
