% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consistency Functions.R
\name{decile_classification2}
\alias{decile_classification2}
\title{Decile classification}
\usage{
decile_classification2(x)
}
\description{
uses exponential decay function with standard deviation of classification score
to evalute consistency. If consistency is < 36, then there are at least a few deciles that have a trend
strongly counter to the general trend of the medical event.
}
