% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{get_event_id}
\alias{get_event_id}
\title{This function returns the Athena event analysis id
affiliated with a particular particular event.}
\usage{
get_event_id(event_type)
}
\arguments{
\item{event_type}{a string-based expression for the medical event of interest
do not include spaces; not case sensitive}
}
\value{
integer analysis id from Athena
}
\description{
This function returns the Athena event analysis id
affiliated with a particular particular event.
}
