% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{get_omop_event_name}
\alias{get_omop_event_name}
\title{This function returns the Athena event analysis name
affiliated with a particular Athena event id.}
\usage{
get_omop_event_name(event_type)
}
\arguments{
\item{event_type}{a string-based expression for the medical event of interest
do not include spaces; not case sensitive}
}
\value{
integer analysis id from Athena
}
\description{
This function returns the Athena event analysis name
affiliated with a particular Athena event id.
}
