% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{group_all_items}
\alias{group_all_items}
\title{This funciton groups an all_items file (a file that has all the classified
items from multiple databases) by a specified knowledge base. First, it condenses
all the classified deciles, so there is 1 entry per concept. The score2 valus
are summed together, so that each concept has 1 value for it's overall trend.
Max value for this parameter is 90 (3 databases * mas score2 of 3 * 10 decades)}
\usage{
group_all_items(item, kb_name, item_file, kb_file, dest_dir)
}
\arguments{
\item{item}{name of the item to be classified}

\item{kb_name}{name of the knowledge based use}

\item{item_file}{full path to the all_item.csv file to group}

\item{kb_file}{full path to the knowledge base file to group by}

\item{dest_dir}{folder where the output .tsv file will go}
}
\value{
none
}
\description{
This funciton groups an all_items file (a file that has all the classified
items from multiple databases) by a specified knowledge base. First, it condenses
all the classified deciles, so there is 1 entry per concept. The score2 valus
are summed together, so that each concept has 1 value for it's overall trend.
Max value for this parameter is 90 (3 databases * mas score2 of 3 * 10 decades)
}
\note{
(\code{c(item, kb_name)}) are only use for naming the output file
}
