% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{plot_db_by_decile}
\alias{plot_db_by_decile}
\title{This function enhances the plot_facet to enable grouping by decile and database
For analyzing all items across all DBs (especially when grouping) this function is best}
\usage{
plot_db_by_decile(df = data.frame())
}
\arguments{
\item{df}{input data.frame; all data from all dbs for ONE concept_id.}
}
\description{
This function enhances the plot_facet to enable grouping by decile and database
For analyzing all items across all DBs (especially when grouping) this function is best
}
\note{
colnames of df must have stratum_1 for concept_id number, stratum_2
for yers, and count_value for prevalence per 1000. edit the function if you want
to have different column names for your variables.
}
