% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consistency Functions.R
\name{qualitative_consistency}
\alias{qualitative_consistency}
\title{@description Qualitative consistency. This calculates the percent difference between
the number of deciles that are rising and sinking. 100 means all deciles are rising or sinking.
0 means there as equal numbers of rising and sinking deciles.}
\usage{
qualitative_consistency(x, n)
}
\arguments{
\item{x}{score}

\item{n}{number of deciles}
}
\description{
@description Qualitative consistency. This calculates the percent difference between
the number of deciles that are rising and sinking. 100 means all deciles are rising or sinking.
0 means there as equal numbers of rising and sinking deciles.
}
