% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{return_rollup_1}
\alias{return_rollup_1}
\title{Rollups}
\usage{
return_rollup_1(databases, event_type, full_cids)
}
\arguments{
\item{databases}{number of databases (user-input at top-level)}

\item{event_type}{type of}

\item{full_cids}{get from lin_list$full_cids (inernal use)}
}
\description{
Rollups
}
\note{
Get df that contains all classified items for a particular event type
}
