% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{rollup_1.2}
\alias{rollup_1.2}
\title{This function summarizes the classified CIDs by counting the number of age-decades
that are rising, strongly rising, etc. for each event type across all DBs (i.e. one row
per event_type account for classification of data in every database)}
\usage{
rollup_1.2(full_cids, databases = "unknown", events, dest_path = "")
}
\arguments{
\item{full_cids}{all classified concept IDs}

\item{databases}{character vector containing the names of the databases}

\item{events}{character vector containing the type of medical event to be analyzed (e.g. drugExposure)}

\item{dest_path}{destination folder for the .tsv file}
}
\description{
This function summarizes the classified CIDs by counting the number of age-decades
that are rising, strongly rising, etc. for each event type across all DBs (i.e. one row
per event_type account for classification of data in every database)
}
