% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centralized Processing Functions.R
\name{sepExportFiles}
\alias{sepExportFiles}
\title{Sep Exports}
\usage{
sepExportFiles(files, descriptor, descriptor_index = 3, delim = "_")
}
\arguments{
\item{descriptor}{Description value (case sensitive)}

\item{descriptor_index}{word position in export file name}

\item{delim}{deliminator in export file name (usually '_')}
}
\description{
get the files that match the descriptor (e.g.3rd word in export
file naming convention)
}
