% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Special Pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Convert \code{data.frame} to \code{list}.}
\arguments{
\item{x}{A \code{data.frame} object.}
}
\description{
Convert \code{data.frame} to \code{list}.
}
\examples{

}
