% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verified Functions.R
\name{step_2}
\alias{step_2}
\title{Step 1 was to pre-process the data. This will read in the data files that
were extracted from the db}
\usage{
step_2(event, pop, event_type = 704, OMOP = FALSE, concept_file = NULL,
  dates)
}
\arguments{
\item{event_type}{whatever naming convention Dr. Huser used to get generate
names for the raw data files. For now, use 3 digit code e.g. 404 for for conditions}

\item{data_folder}{where the .csv files produced by step1 live}
}
\value{
pre-processed eventM
}
\description{
Step 2 Preprocess the raw data extracted from the databse.
    This involves joining the original data with the Athena concept tables
    to get concept_names.
}
